cpp-defaults: delete and default the appropriate defaults
The defaults supplied by the compiler always ease the task
and spread an invisible veil over what's going on in the
background.

The background scenes by trying to delete all the defaults
provided which revealed which default was being used at what
places. As a consequence two empty but required constructors
were defaulted. Exact usage of a few of the defaults can be
found in the comments.

It was interesting to hunt these down! :smile: