hill_cipher: normalize the matrix of int values
When the plain text is converted to corresponding integer values
the number of values might not be a multiple of 'degree' which
is required for the correct working of the algorithm.

Pad the remaining places, if any, with zeros to make sure the number
of integers is a multiple of degree.