playfair_cipher: fix REGRESSION as result of REVERT
This teaches/re-inforces the fact that a change needs to include
only what it "says it does". Of course, it implies that 'changes
are independent'.

This has also re-inforced the fact that the header needed by an
implementation need to be declared 'explicitly' instead of expecting
them to be included in some other file.