pipe-joiner: better check the input values
Improve the user interface by checking for failure and
asking input until input is read succesfully.

Further, in case the solution is trying to be printed even before
the problem is solved show a warning. This would happen when the
solution is tryin to be printed before a success input read.