pipe-joiner: make the computation async
This invokes the core function asynchronously and gets the
value only when the output is required. The function may have
executed any time between the time at which the async task was
created and the time at which the results is requested.

A check needs to be done to identify if the async task was started
in the first place.