railfence: add initial (ad-hoc) implementation
The implementation was ad-hoc in the sense that it's clumsy
by doing the same thing repeatedly in several places. It has
to be cleaned up.

The process of cleaning it up is to be done as a series of commits
which could serve as an enlightening history of correcting mistakes.