railfence: optimize fill_matrix
The unwanted logic to travers the matrix in different ways
(row-wise/column-wise) w.r.t a flag has been removed. That's
because it's much easier to access the string differently
w.r.t the flag.

This reduces unwanted cost of two extra variables which also
confused things a little. The cost of accessing string is same
as was before as only the order in which elements are accessed
differ.